chore: unify components
