--- frontend/components/RecommendationsComponent.tsx
@@ -1,8 +1,0 @@
-export default function RecommendationsComponent() {
-    return (
-      <div>
-        <h2 className="text-2xl font-bold mb-4">Recommendations</h2>
-        <p>This is where movie recommendations will be shown.</p>
-      </div>
-    );
-  }
